requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.2.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
audioread==2.1.9
autoclasstoc==1.5.1
autograd==1.3
Babel==2.10.3
backcall==0.2.0
binaryornot==0.4.4
bleach==4.0.0
blessings==1.7
cached-property==1.5.2
cachetools==4.2.2
catboost==1.0.6
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
cloudpickle==1.6.0
cma==2.7.0
colorama==0.4.4
colorful==0.5.4
ConfigSpace==0.4.19
cookiecutter==1.7.3
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
descartes==1.1.0
dill==0.3.4
dm-tree==0.1.6
docopt==0.6.2
docutils==0.17.1
edward2 @ git+https://github.com/google/edward2.git@194a9841235126219743bc52a2e36d48f0f35965
entrypoints==0.3
filelock==3.0.12
fire==0.4.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin-config==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
GPy==1.10.0
gpytorch==1.5.0
graphviz==0.20
grpcio==1.39.0
gym==0.18.3
h5py==3.1.0
hiredis==2.0.0
hpbandster==0.7.4
httplib2==0.19.1
idna==3.2
imagesize==1.4.1
importlib-metadata==4.12.0
importlib-resources==5.2.2
ipykernel==6.15.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jinja2-time==0.2.0
joblib==1.0.1
jsonpickle==1.5.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.12
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.38.1
m2r2==0.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
more-itertools==8.13.0
msgpack==1.0.2
multidict==5.1.0
munch==2.5.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.11.0
notebook==6.4.0
numba==0.55.2
nuscenes-devkit==1.1.5
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
opencv-python==4.5.1.48
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
paramz==0.9.5
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip-licenses==3.5.4
plotly==5.9.0
pooch==1.6.0
portalocker==2.3.0
poyo==0.5.0
prettytable==2.1.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py-cpuinfo==8.0.0
py-spy==0.3.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydantic==1.8.2
pyglet==1.5.15
Pygments==2.9.0
pymoo==0.5.0
pyparsing==2.4.7
pyquaternion==0.9.9
Pyro4==4.80
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
ray==1.5.1
redis==3.5.3
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.3.0
rsa==4.7.2
sacrebleu==2.0.0
sacred==0.8.2
scikit-learn==0.24.1
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.7.1
sentencepiece==0.1.96
seqeval==1.2.2
serpent==1.40
Shapely==1.7.1
six==1.15.0
smmap==4.0.0
#snowballstemmer==2.2.0
SoundFile==0.10.3.post1
Sphinx==5.0.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-confluencebuilder==1.8.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
statsmodels==0.12.2
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==2.6.0
tensorflow-addons==0.13.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.2.0
tensorflow-model-optimization==0.6.0
tensorflow-probability==0.14.1
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
tf-models-official==2.5.1
tf-slim==1.1.0
threadpoolctl==2.2.0
tornado==6.1
tqdm==4.60.0
traitlets==5.3.0
typeguard==2.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.5.0

